requirements.txt
Flask==2.1.0
Flask_Cors==3.0.10
python-dotenv==0.20.0
requests==2.28.0
